refactor index page, fix mobile styles to match design
