+++ __tests__/index.test.jsx
@@ -1,6 +1,9 @@
 import '@testing-library/jest-dom'
 import { render, screen } from '@testing-library/react'
+import mockRouter from 'next-router-mock';
 import Home from '../src/pages/index';
+
+jest.mock('next/router', () => jest.requireActual('next-router-mock'))
 
 describe('Home', () => {
   it('renders a heading', () => {
