+++ package.json
@@ -19,10 +19,12 @@
   "dependencies": {
     "classnames": "^2.5.1",
     "next": "14.1.0",
+    "next-router-mock": "^0.9.12",
     "react": "^18",
     "react-device-detect": "^2.2.3",
     "react-dom": "^18",
     "react-use": "^17.5.0",
+    "save-dev": "^0.0.1-security",
     "swr": "^2.2.4"
   },
   "devDependencies": {
