+++ src/components/Control/stylesheet/size/main.module.scss
@@ -10,6 +10,10 @@
       'medium': (
         'width': 296px,
         'height': 64px
+      ),
+      'responsive': (
+        'width': 100%,
+        'height': 100%
       )
     );
