--- src/styles/final-page.scss
@@ -1,4 +1,0 @@
-.final-block {
-  background: #212531;
-  background: linear-gradient(to right bottom, #f5f5f7 50%, #fff3eb 50%);
-}
